Fix status codes and add videoUrl in Like

+++ routes/songs.js
@@ -33,8 +33,8 @@ router.get("/:id", async function (req, res, next) {
         user: 1,
         date: 1,
       });
-      if (result?.length > 0) res.send(result);
-      else res.sendStatus(204);
+      if (result) res.send(result);
+      else throw new Error("Invalid song");
     } else {
       next();
     }
@@ -79,7 +79,7 @@ router.get("/spotify", async function (req, res, next) {
             artists: [],
             albumCover: track.album.images[0].url,
             releaseDate: track.album.release_date,
-            url: track.preview_url,
+            audioUrl: track.preview_url,
             spotifyId: track.id,
           };
           track.artists.forEach((artist) => {
@@ -101,23 +101,23 @@ router.get("/spotify", async function (req, res, next) {
 
 router.post("/", async function (req, res, next) {
   try {
-    const { title, artists, releaseDate, albumCover, url, lyrics, spotifyId } =
+    const { title, artists, releaseDate, albumCover, audioUrl, videoUrl, lyrics, spotifyId } =
       req.body;
     const song = new Song({
       title,
       artists,
       releaseDate: new Date(releaseDate),
       albumCover,
-      url,
+      audioUrl: audioUrl,
+      videoUrl,
       lyrics,
       spotifyId,
       likes: [],
     });
     await song.save();
-    return res.sendStatus(201);
+    return res.status(201).send(song);
   } catch (err) {
-    if (err?.code === 11000) res.status(409).send("Conflict: Duplicate");
-    else next(err);
+    next(err);
   }
 });
 
@@ -143,14 +143,15 @@ router.post("/ticket", async function (req, res, next) {
 
 router.put("/", async function (req, res, next) {
   try {
-    const { id, url, lyrics } = req.body;
-    if (Object.keys(req.query).length === 0) {
+    const { id, url, videoUrl, lyrics } = req.body;
+    if (Object.keys(req.body).length >= 2) {
       const newInfo = {};
-      if (typeof url !== "undefined") newInfo.url = url;
+      if (typeof url !== "undefined") newInfo.videoUrl = url;
+      if (typeof videoUrl !== "undefined") newInfo.videoUrl = videoUrl;
       if (typeof lyrics !== "undefined") newInfo.lyrics = lyrics;
 
       const result = await Song.findByIdAndUpdate(id, newInfo, {
-        new: true,
+        new: true
       }).populate("likes", {
         user: 1,
         date: 1,
@@ -168,8 +169,12 @@ router.delete("/:id", async function (req, res, next) {
   try {
     const id = req.params.id;
     const result = await Song.findByIdAndDelete(id);
-    const deletedLikes = await Like.deleteMany({ song: id });
-    res.sendStatus(204);
+    if (result) {
+      await Like.deleteMany({ song: id });
+      res.sendStatus(204);
+    } else {
+      throw new Error("Invalid song");
+    }
   } catch (err) {
     next(err);
   }
